feat: moved banner to organisms
