+++ crowdin/messages.json
@@ -1785,6 +1785,8 @@
   "-13898976": "It's been a really great experience trading forex on Deriv - it's a smooth and seamless operation!",
   "-143741607": "Deriv is the best forex broker I have ever come across!",
   "-1026502273": "One broker,",
+  "-853506275": "Make hassle-free deposits and withdrawals today",
+  "-1958137973": "Android, iOS, and Huawei",
   "-402834354": "Google Play",
   "-1866414327": "App Store",
   "-604256139": "AppGallery",
@@ -2236,8 +2238,6 @@
   "-231290085": "Start connecting with fellow traders to make deposits and withdrawals.",
   "-1719881739": "On your mobile",
   "-399746134": "How to get Deriv P2P",
-  "-1958137973": "Android, iOS, and Huawei",
-  "-853506275": "Make hassle-free deposits and withdrawals today",
   "-385361251": "Save time",
   "-1139543787": "Works with your local currency",
   "-1267106447": "Deal with fellow traders at pre-agreed rates.",
