+++ src/features/components/organisms/banner/blade-shape-banner/styles.module.scss
