+++ src/features/pages/p2p/index.tsx
@@ -5,8 +5,9 @@ import DP2Numbers from './numbers'
 import DP2P from './dp2p'
 import DP2Steps from './steps'
 import DP2Availability from './availability'
-import DP2Banner from './banner'
 import Roadmap from './roadmap'
+import { banner_data } from './_data'
+import DP2Banner from 'features/components/organisms/banner/blade-shape-banner'
 import { useOpenLiveChat } from 'components/hooks/use-open-live-chat-redirection'
 import Layout from 'features/components/templates/layout'
 import { SEO } from 'components/containers'
@@ -39,7 +40,7 @@ const DP2PPage = () => {
                 <DP2P />
                 <DP2Steps />
                 <DP2Availability />
-                <DP2Banner />
+                <DP2Banner data={banner_data} />
                 <Roadmap />
                 <Footer />
             </Layout>
